trading_bot/requirements.txt
# Velocity + Acceleration Trading Bot
# Python package requirements

# Alpaca Trading API
alpaca-trade-api>=3.0.0

# Data manipulation
pandas>=2.0.0
numpy>=1.24.0

# Environment variables
python-dotenv>=1.0.0

# Technical analysis (optional, for future enhancements)
ta-lib>=0.4.0

# Data visualization (optional, for analysis)
matplotlib>=3.7.0
seaborn>=0.12.0

# Progress bars (optional)
tqdm>=4.65.0

# Requests for API calls
requests>=2.31.0
